monday-bi-agent/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
requests==2.32.4
pandas==2.3.2
python-dateutil==2.9.0.post0
openai==1.107.1
pydantic==2.11.7
pydantic-settings==2.10.1
numpy==2.3.2